Update readme to make it clear manual proguard configuration should be unnecessary  (#69)
For some time we've supported consumer Proguard rules which are included in the aar library distributions, and automatically bundled into Proguard configurations for client applications. The readme now reflects this, but includes a link to the up to date Proguard configuration if manual configuration is needed. It also includes mention of the more recent Android specific code shrinking tool R8. Thanks to #68 for the suggestion.