Do not syncrhonize Throttler::run
Throttler::run and Throttler::attemptRun are the two functions that can syncrhonize both itself and
ConnectivityManager in that order, ConnectivityManager itself usually syncrhonizes on
itself then on Throttler by calling its functions. Because Throttler::run is run on a separate
thread this becomes a problem as the order of locking is different. Removing synchronize
from Throttler::run will restore the order as the runnable will lock ConnectivityManager.
Throttler::attemptRun is only called from ConnectivityManager and after it locks ConnectivityManager
so Throttler::attemptRun should not require any changes.