fix: add the result callback for ping stream mode (#296)
**Requirements**

- [x] I have added test coverage for new or changed functionality
- [x] I have followed the repository's [pull request submission
guidelines](../blob/main/CONTRIBUTING.md#submitting-pull-requests)
- [x] I have validated my changes against all supported platform
versions

**Related issues**

https://app.shortcut.com/launchdarkly/story/205006/escalation-customer-found-that-after-upgrading-the-rn-sdk-from-6-2-4-to-7-1-4-our-android-app-s-ld-integration-started-failing

**Describe the solution you've provided**
Because of the rework of the SDK streaming data source module, we missed
a change to signal to the SDK that initialization is complete, for the
ping and fetch case (which uses by Relay). This callback happens for
both Android 3.x and iOS currently, so we are adding this back to
Android 4.x.

Issue: The method used are private and hard to test individually.

**Describe alternatives you've considered**
Potentially making some method public, or making a public testing
wrapper of a method that we don't expose. But that is a janky pattern so
I want to discuss.

**Additional context**
N/A