fix: synchronize ConnectivityManager.updateDataSource to prevent race… (#311)
… condition

**Requirements**

- [ ] I have added test coverage for new or changed functionality
- [x] I have followed the repository's [pull request submission
guidelines](../blob/main/CONTRIBUTING.md#submitting-pull-requests)
- [ ] I have validated my changes against all supported platform
versions

**Related issues**


https://app.shortcut.com/launchdarkly/story/229919/support-58536-crash-initializing-android-sdk

**Describe the solution you've provided**

Root cause of crash is that multiple calls to update from different
sources can cause a event source to be stopped just before start is
called on the event source. This results in the event source being in a
bad state, leading to the crash. The fix is to only allow one thread
into the updateDataSource block at a time by making it synchronized.

**Describe alternatives you've considered**

Also considered updating the event source to silently ignore the request
to start again after it has been stopped. I prefer code less
permissive/lenient code in this case since we really should know exactly
when we are starting and stopping eventsources.