fix: correcting protocol parsing logic that may lead to incorrect sanitization of an incoming message from LD servers (#288)
After examining related code to a customer support case and #278,
noticed this modification during iteration. Theoretical issue is that
modification during iteration, even if not resulting in
ConcurrentModificationExceptions, could result in unpredictable
iteration and could possibly skip sanitizing certain incoming flags.

Through bench testing and instrumentation testing, I was unable to
reproduce the issue, but I suspect it may be dependent on the specific
platform's implementation of the Map type that GSON will depend on. Will
deploy this fix and work with customers to confirm if the occurrences of
the issue is eliminated.